Wrap ExportDialog.js api
